Rename go-homedir package to homedir.
Simplifies the package name by removing the "go-" prefix, making it
more concise while maintaining the same functionality. The package
was already internally named "homedir" so this aligns the directory
name with the actual package name.
