+++ pkg/config/homedir/ATTRIBUTIONS.md
