+++ pkg/config/homedir/README.md
